allow non-CLI variable assignments
If we don't want to specify $usernames and $startdate on the command line,
but instead just want to assign the relevant variables at the top of the
script file, this will allow that.  Passing the arguments on the CLI
will override any saved variables in the script, as is proper.